requirements.txt
aiohttp==3.8.1
dagit==1.0.7
dagster-postgres==0.16.7
dagster==1.0.7
pandas==1.4.4
psycopg2-binary==2.9.3
SQLAlchemy==1.4.41
urllib3==1.26.12